benchmark

+++ src/algorithms/edit_distance.h
@@ -28,19 +28,19 @@ private:
 
 public:
     // Edit distance on real sequences
-    static float edr(Vec3D *v1, Vec3D *v2, int size_v1, int size_v2, float epsilon,
-                     std::function<float(const Vec3D &, const Vec3D &)> func);
+    static int edr(Vec3D *v1, Vec3D *v2, int size_v1, int size_v2, float epsilon,
+                   std::function<float(const Vec3D &, const Vec3D &)> func);
 
-    static float edr(const std::vector<Vec3D *> &inpF, const std::vector<Vec3D *> &refF, float epsilon,
-                     std::function<float(const Vec3D *, const Vec3D *)> func);
+    static int edr(const std::vector<Vec3D *> &inpF, const std::vector<Vec3D *> &refF, float epsilon,
+                   std::function<float(const Vec3D *, const Vec3D *)> func);
 
-    static float edr(const std::vector<Quaternion *> &inpF, const std::vector<Quaternion *> &refF, float epsilon,
-                     std::function<float(const Quaternion *, const Quaternion *,
-                                         const std::array<bool, JOINT_COUNT> &selectedJ)> &func);
+    static int edr(const std::vector<Quaternion *> &inpF, const std::vector<Quaternion *> &refF, float epsilon,
+                   std::function<float(const Quaternion *, const Quaternion *,
+                                       const std::array<bool, JOINT_COUNT> &selectedJ)> &func);
 
-    static float edr(const std::vector<Quaternion *> &inpF, const std::vector<Quaternion *> &refF, float epsilon,
-                     std::function<float(const Quaternion *, const Quaternion *,
-                                         const std::array<float, JOINT_COUNT> &selectedJ)> &func);
+    static int edr(const std::vector<Quaternion *> &inpF, const std::vector<Quaternion *> &refF, float epsilon,
+                   std::function<float(const Quaternion *, const Quaternion *,
+                                       const std::array<float, JOINT_COUNT> &selectedJ)> &func);
 
     // lambda := Penalty for deletion operation
     // nu     := Elasticity parameter - nu >=0 needed for distance measure
